+++ src/worker/worker.ts
@@ -35,13 +35,6 @@ export interface SessionWorkerOptions {
   };
   runner: OpencodeRunner;
   logger: Logger;
-  limits?: {
-    groupWindowEntries?: number;
-    userMemoryEntries?: number;
-    /** Deprecated: use groupWindowEntries/userMemoryEntries. */
-    historyEntries?: number;
-    historyBytes?: number;
-  };
 }
 
 export class SessionWorker {
@@ -90,17 +83,6 @@ export class SessionWorker {
       runner: options.runner,
       activityIndex,
       bufferStore,
-      limits: {
-        groupWindowEntries:
-          options.limits?.groupWindowEntries ??
-          options.limits?.historyEntries ??
-          30,
-        userMemoryEntries:
-          options.limits?.userMemoryEntries ??
-          options.limits?.historyEntries ??
-          20,
-        historyBytes: options.limits?.historyBytes ?? 200_000,
-      },
     });
 
     this.worker = new Worker<SessionJobData>(